Misha, we have TWO NetCDF readers. Hence, double -> float in second one too
git-svn-id: svn://bar.ebi.ac.uk/branches/atlas-standalone@11551 2913f559-6b04-0410-9a09-c530ee9f5186

+++ atlas-web/src/main/java/ae3/dao/NetCDFReader.java
@@ -203,8 +203,8 @@ public class NetCDFReader {
                         int[] originBDC = new int[varPVAL.getRank()];
                         originBDC[0] = designElementId;
                         shapeBDC[0] = 1;
-                        double[] pvals = (double[])varPVAL.read(originBDC, shapeBDC).reduce().get1DJavaArray(double.class);
-                        double[] tstats = (double[])varTSTAT.read(originBDC, shapeBDC).reduce().get1DJavaArray(double.class);
+                        float[] pvals = (float[])varPVAL.read(originBDC, shapeBDC).reduce().get1DJavaArray(float.class);
+                        float[] tstats = (float[])varTSTAT.read(originBDC, shapeBDC).reduce().get1DJavaArray(float.class);
 
                         EfvTree<Stat> result = new EfvTree<Stat>();
                         for(EfvTree.EfEfv<Integer> efefv : efvTree.getNameSortedList()) {
